DAOS-2955 test: security primary group user ACL verification                 and DAOS-2951, 2957, 2961, 2962, 2963.
new file:     src/tests/ftest/security/pool_security_groups.py
new file:     src/tests/ftest/security/pool_security_groups.yaml
new file:     src/tests/ftest/util/pool_security_test_base.py
updated file: src/tests/ftest/security/pool_security_acl.py
updated file: src/tests/ftest/security/pool_security_acl.yaml

Test-tag: pr,-hw security,pool_acl
