+++ src/tests/ftest/security/pool_security_groups.yaml
@@ -1,0 +1,82 @@
+# change host names to your reserved nodes, the
+# required quantity is indicated by the placeholders
+hosts:
+  test_servers:
+    - server-A
+  test_clients:
+    - server-A
+  transport_config:
+    - allow_insecure: True
+    - ca_cert: ~/.daos/daosCA.crt
+    - cert: ~/.daos/daos_agent.crt
+    - key: ~/.daos/daos_agent.key
+    - server_name: server
+timeout: 600
+server_config:
+  name: daos_server
+  port: 10001
+daos_server:
+  start:
+     insecure: True
+dmg:
+  request: pool
+  action: create
+  create:
+    scm_size: 16000000
+pool_acl:
+  user_prefix: daos_ci
+  num_user: 1
+  num_group: 1
+  primary_secondary_group_test:
+    primary_group:
+      pg_permission:
+        - rw
+      pg_read_write:
+        - pass
+        - pass
+    secondary_group_name:
+      - daos_ci_sg2
+      - daos_ci_sg3
+    test_case: !mux
+      test1:
+        sg_permission:
+          - r
+          - w
+        sg_read_write:
+          - pass
+          - pass
+      test2:
+        sg_permission:
+          - none
+          - rw
+        sg_read_write:
+          - pass
+          - pass
+      test3:
+        sg_permission:
+          - r
+          - none
+        sg_read_write:
+          - pass
+          - deny
+      test4:
+        sg_permission:
+          - none
+          - w
+        sg_read_write:
+          - deny
+          - deny
+      test5:
+        sg_permission:
+          - wr
+          - w
+        sg_read_write:
+          - pass
+          - pass
+      test6:
+        sg_permission:
+          - none
+          - none
+        sg_read_write:
+          - deny
+          - deny
